+++ demos/example-vite/vite.config.ts
@@ -13,17 +13,23 @@ export default defineConfig({
     },
     emptyOutDir: true
   },
+  resolve: {
+    alias: [
+      // https://jira.mongodb.org/browse/NODE-5773
+      { find: 'bson', replacement: require.resolve('bson') }
+    ]
+  },
   envDir: '..', // Use this dir for env vars, not 'src'.
   optimizeDeps: {
     // Don't optimize these packages as they contain web workers and WASM files.
     // https://github.com/vitejs/vite/issues/11672#issuecomment-1415820673
     exclude: ['@journeyapps/wa-sqlite', '@powersync/web'],
     include: [
-      '@powersync/web > uuid',
       '@powersync/web > event-iterator',
       '@powersync/web > js-logger',
       '@powersync/web > lodash/throttle',
-      '@powersync/web > can-ndjson-stream'
+      '@powersync/web > can-ndjson-stream',
+      '@powersync/web > bson'
     ]
   },
   plugins: [wasm(), topLevelAwait()],