+++ demos/react-native-supabase-todolist/app/views/todos/_layout.tsx
@@ -1,5 +1,4 @@
 import { Stack } from 'expo-router';
-import React from 'react';
 
 /**
  * Stack to edit a list from the todos view