+++ packages/common/src/client/AbstractPowerSyncDatabase.ts
@@ -25,7 +25,8 @@ import {
   AbstractStreamingSyncImplementation,
   DEFAULT_CRUD_UPLOAD_THROTTLE_MS,
   StreamingSyncImplementationListener,
-  StreamingSyncImplementation
+  StreamingSyncImplementation,
+  PowerSyncConnectionOptions
 } from './sync/stream/AbstractStreamingSyncImplementation';
 
 export interface DisconnectAndClearOptions {
@@ -313,7 +314,7 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
   /**
    * Connects to stream of events from the PowerSync instance.
    */
-  async connect(connector: PowerSyncBackendConnector) {
+  async connect(connector: PowerSyncBackendConnector, options?: PowerSyncConnectionOptions) {
     await this.waitForReady();
 
     // close connection if one is open
@@ -325,14 +326,17 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
     this.syncStreamImplementation = this.generateSyncStreamImplementation(connector);
     this.syncStatusListenerDisposer = this.syncStreamImplementation.registerListener({
       statusChanged: (status) => {
-        this.currentStatus = new SyncStatus({ ...status.toJSON(), hasSynced: this.currentStatus?.hasSynced });
+        this.currentStatus = new SyncStatus({
+          ...status.toJSON(),
+          hasSynced: this.currentStatus?.hasSynced || !!status.lastSyncedAt
+        });
         this.iterateListeners((cb) => cb.statusChanged?.(this.currentStatus));
       }
     });
 
     await this.syncStreamImplementation.waitForReady();
     this.syncStreamImplementation.triggerCrudUpload();
-    await this.syncStreamImplementation.connect();
+    await this.syncStreamImplementation.connect(options);
   }
 
   /**
@@ -766,7 +770,7 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
         `SELECT DISTINCT tbl_name FROM sqlite_master WHERE rootpage IN (SELECT json_each.value FROM json_each(?))`,
         [JSON.stringify(rootPages)]
       );
-      for (let table of tables) {
+      for (const table of tables) {
         resolvedTables.push(table.tbl_name.replace(POWERSYNC_TABLE_MATCH, ''));
       }
     }
@@ -937,7 +941,7 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
       ? filteredTables
       : filteredTables.map((t) => t.replace(POWERSYNC_TABLE_MATCH, ''));
 
-    for (let table of mappedTableNames) {
+    for (const table of mappedTableNames) {
       changedTables.add(table);
     }
   }