+++ packages/common/src/client/sync/stream/AbstractStreamingSyncImplementation.ts
@@ -4,16 +4,13 @@ import Logger, { ILogger } from 'js-logger';
 
 import {
   BucketRequest,
-  StreamingSyncLine,
-  StreamingSyncRequest,
   isStreamingKeepalive,
   isStreamingSyncCheckpoint,
   isStreamingSyncCheckpointComplete,
   isStreamingSyncCheckpointDiff,
   isStreamingSyncData
 } from './streaming-sync-types';
-import { AbstractRemote } from './AbstractRemote';
-import ndjsonStream from 'can-ndjson-stream';
+import { AbstractRemote, SyncStreamOptions } from './AbstractRemote';
 import { BucketChecksum, BucketStorageAdapter, Checkpoint } from '../bucket/BucketStorageAdapter';
 import { SyncStatus, SyncStatusOptions } from '../../../db/crud/SyncStatus';
 import { SyncDataBucket } from '../bucket/SyncDataBucket';
@@ -24,6 +21,12 @@ export enum LockType {
   CRUD = 'crud',
   SYNC = 'sync'
 }
+
+export enum SyncStreamConnectionMethod {
+  HTTP = 'http',
+  WEB_SOCKET = 'web-socket'
+}
+
 /**
  * Abstract Lock to be implemented by various JS environments
  */
@@ -59,11 +62,24 @@ export interface StreamingSyncImplementationListener extends BaseListener {
   statusChanged?: ((status: SyncStatus) => void) | undefined;
 }
 
+/**
+ * Configurable options to be used when connecting to the PowerSync
+ * backend instance.
+ */
+export interface PowerSyncConnectionOptions {
+  /**
+   * The connection method to use when streaming updates from
+   * the PowerSync backend instance.
+   * Defaults to a HTTP streaming connection.
+   */
+  connectionMethod?: SyncStreamConnectionMethod;
+}
+
 export interface StreamingSyncImplementation extends BaseObserver<StreamingSyncImplementationListener>, Disposable {
   /**
    * Connects to the sync service
    */
-  connect(): Promise<void>;
+  connect(options?: PowerSyncConnectionOptions): Promise<void>;
   /**
    * Disconnects from the sync services.
    * @throws if not connected or if abort is not controlled internally
@@ -87,6 +103,10 @@ export const DEFAULT_STREAMING_SYNC_OPTIONS = {
   crudUploadThrottleMs: DEFAULT_CRUD_UPLOAD_THROTTLE_MS
 };
 
+export const DEFAULT_STREAM_CONNECTION_OPTIONS: Required<PowerSyncConnectionOptions> = {
+  connectionMethod: SyncStreamConnectionMethod.HTTP
+};
+
 export abstract class AbstractStreamingSyncImplementation
   extends BaseObserver<StreamingSyncImplementationListener>
   implements StreamingSyncImplementation
@@ -103,6 +123,7 @@ export abstract class AbstractStreamingSyncImplementation
   constructor(options: AbstractStreamingSyncImplementationOptions) {
     super();
     this.options = { ...DEFAULT_STREAMING_SYNC_OPTIONS, ...options };
+
     this.syncStatus = new SyncStatus({
       connected: false,
       lastSyncedAt: undefined,
@@ -230,13 +251,13 @@ export abstract class AbstractStreamingSyncImplementation
     }
   }
 
-  async connect() {
+  async connect(options?: PowerSyncConnectionOptions) {
     if (this.abortController) {
       await this.disconnect();
     }
 
     this.abortController = new AbortController();
-    this.streamingSyncPromise = this.streamingSync(this.abortController.signal);
+    this.streamingSyncPromise = this.streamingSync(this.abortController.signal, options);
 
     // Return a promise that resolves when the connection status is updated
     return new Promise<void>((resolve) => {
@@ -288,7 +309,7 @@ export abstract class AbstractStreamingSyncImplementation
   /**
    * @deprecated use [connect instead]
    */
-  async streamingSync(signal?: AbortSignal): Promise<void> {
+  async streamingSync(signal?: AbortSignal, options?: PowerSyncConnectionOptions): Promise<void> {
     if (!signal) {
       this.abortController = new AbortController();
       signal = this.abortController.signal;
@@ -334,7 +355,7 @@ export abstract class AbstractStreamingSyncImplementation
         if (signal?.aborted) {
           break;
         }
-        const { retry } = await this.streamingSyncIteration(nestedAbortController.signal);
+        const { retry } = await this.streamingSyncIteration(nestedAbortController.signal, options);
         if (!retry) {
           /**
            * A sync error ocurred that we cannot recover from here.
@@ -377,11 +398,19 @@ export abstract class AbstractStreamingSyncImplementation
     this.updateSyncStatus({ connected: false });
   }
 
-  protected async streamingSyncIteration(signal: AbortSignal, progress?: () => void): Promise<{ retry?: boolean }> {
+  protected async streamingSyncIteration(
+    signal: AbortSignal,
+    options?: PowerSyncConnectionOptions
+  ): Promise<{ retry?: boolean }> {
     return await this.obtainLock({
       type: LockType.SYNC,
       signal,
       callback: async () => {
+        const resolvedOptions: Required<PowerSyncConnectionOptions> = {
+          ...DEFAULT_STREAM_CONNECTION_OPTIONS,
+          ...(options ?? {})
+        };
+
         this.logger.debug('Streaming sync iteration started');
         this.options.adapter.startSession();
         const bucketEntries = await this.options.adapter.getBucketStates();
@@ -403,14 +432,40 @@ export abstract class AbstractStreamingSyncImplementation
 
         let bucketSet = new Set<string>(initialBuckets.keys());
 
-        for await (const line of this.streamingSyncRequest(
-          {
+        this.logger.debug('Requesting stream from server');
+
+        const syncOptions: SyncStreamOptions = {
+          path: '/sync/stream',
+          abortSignal: signal,
+          data: {
             buckets: req,
             include_checksum: true,
             raw_data: true
-          },
-          signal
-        )) {
+          }
+        };
+
+        const stream =
+          resolvedOptions?.connectionMethod == SyncStreamConnectionMethod.HTTP
+            ? await this.options.remote.postStream(syncOptions)
+            : await this.options.remote.socketStream(syncOptions);
+
+        this.logger.debug('Stream established. Processing events');
+
+        while (!stream.closed) {
+          const line = await stream.read();
+          if (!line) {
+            // The stream has closed while waiting
+            return { retry: true };
+          }
+          // A connection is active and messages are being received
+          if (!this.syncStatus.connected) {
+            // There is a connection now
+            Promise.resolve().then(() => this.triggerCrudUpload());
+            this.updateSyncStatus({
+              connected: true
+            });
+          }
+
           if (isStreamingSyncCheckpoint(line)) {
             targetCheckpoint = line.checkpoint;
             const bucketsToDelete = new Set<string>(bucketSet);
@@ -495,6 +550,11 @@ export abstract class AbstractStreamingSyncImplementation
             if (remaining_seconds == 0) {
               // Connection would be closed automatically right after this
               this.logger.debug('Token expiring; reconnect');
+              /**
+               * For a rare case where the backend connector does not update the token
+               * (uses the same one), this should have some delay.
+               */
+              await this.delayRetry();
               return { retry: true };
             }
             this.triggerCrudUpload();
@@ -528,43 +588,12 @@ export abstract class AbstractStreamingSyncImplementation
               }
             }
           }
-          progress?.();
         }
         this.logger.debug('Stream input empty');
         // Connection closed. Likely due to auth issue.
         return { retry: true };
       }
     });
-  }
-
-  protected async *streamingSyncRequest(
-    req: StreamingSyncRequest,
-    signal?: AbortSignal
-  ): AsyncGenerator<StreamingSyncLine> {
-    const body = await this.options.remote.postStreaming('/sync/stream', req, {}, signal);
-
-    // A connection is active
-    // There is a connection now
-    Promise.resolve().then(() => this.triggerCrudUpload());
-    this.updateSyncStatus({
-      connected: true
-    });
-
-    const stream = ndjsonStream(body);
-    const reader = stream.getReader();
-
-    try {
-      while (true) {
-        // Read from the stream
-        const { done, value } = await reader.read();
-        // Exit if we're done
-        if (done) return;
-        // Else yield the chunk
-        yield value;
-      }
-    } finally {
-      reader.releaseLock();
-    }
   }
 
   protected updateSyncStatus(options: SyncStatusOptions) {