+++ packages/common/src/utils/BaseObserver.ts
@@ -1,5 +1,3 @@
-import { v4 } from 'uuid';
-
 export interface Disposable {
   dispose: () => Promise<void>;
 }
@@ -13,26 +11,29 @@ export type BaseListener = {
 };
 
 export class BaseObserver<T extends BaseListener = BaseListener> implements BaseObserverInterface<T> {
-  protected listeners: { [id: string]: Partial<T> };
+  protected listeners = new Set<Partial<T>>();
 
-  constructor() {
-    this.listeners = {};
-  }
+  constructor() {}
 
   /**
    * Register a listener for updates to the PowerSync client.
    */
   registerListener(listener: Partial<T>): () => void {
-    const id = v4();
-    this.listeners[id] = listener;
+    this.listeners.add(listener);
     return () => {
-      delete this.listeners[id];
+      this.listeners.delete(listener);
     };
   }
 
   iterateListeners(cb: (listener: Partial<T>) => any) {
-    for (const i in this.listeners) {
-      cb(this.listeners[i]);
+    for (const listener of this.listeners) {
+      cb(listener);
+    }
+  }
+
+  async iterateAsyncListeners(cb: (listener: Partial<T>) => Promise<any>) {
+    for (let i of Array.from(this.listeners.values())) {
+      await cb(i);
     }
   }
 }