+++ demos/example-vite/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-vite
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [395ea24]
+- Updated dependencies [9d1dc6f]
+  - @powersync/web@0.6.0
+
+## 0.0.18
+
+### Patch Changes
+
+- @powersync/web@0.5.3
+
 ## 0.0.17
 
 ### Patch Changes