+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/react
 
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [395ea24]
+- Updated dependencies [9d1dc6f]
+  - @powersync/common@1.8.0
+
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [3c421ea]
+  - @powersync/common@1.7.1
+
 ## 1.3.0
 
 ### Minor Changes