+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/vue
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [395ea24]
+- Updated dependencies [9d1dc6f]
+  - @powersync/common@1.8.0
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [3c421ea]
+  - @powersync/common@1.7.1
+
 ## 0.1.0
 
 ### Minor Changes