+++ packages/web/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @powersync/web
 
+## 0.6.0
+
+### Minor Changes
+
+- 9d1dc6f: Added support for WebSocket sync stream connections.
+
+### Patch Changes
+
+- 395ea24: Remove uuid dependency
+- Updated dependencies [395ea24]
+- Updated dependencies [9d1dc6f]
+  - @powersync/common@1.8.0
+
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [3c421ea]
+  - @powersync/common@1.7.1
+
 ## 0.5.2
 
 ### Patch Changes