+++ tools/diagnostics-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # diagnostics-app
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [395ea24]
+- Updated dependencies [9d1dc6f]
+  - @powersync/web@0.6.0
+  - @powersync/react@1.3.2
+
+## 0.1.3
+
+### Patch Changes
+
+- @powersync/react@1.3.1
+- @powersync/web@0.5.3
+
 ## 0.1.2
 
 ### Patch Changes